+++ examples/node/player.js
@@ -24,11 +24,11 @@ operational.to(flipped).when(function (s) { return s === "flip"; });
 flipped.to(operational).when(function (s) { return s === "flip"; });
 operational.to(finalState).when(function (s) { return s === "off"; });
 // create a new state machine instance (this stores the active state configuration, allowing many instances to work with a single model)
-var instance = new state.JSONInstance("player");
+var instance = new state.DictionaryInstance("player");
 // initialse the state machine instance (also initialises the model if not already initialised explicitly or via another instance)
-state.initialise(model, instance);
+model.initialise(instance);
 // send messages to the state machine to cause state transitions
-state.evaluate(model, instance, "play");
-state.evaluate(model, instance, "pause");
-state.evaluate(model, instance, "stop");
-console.log(instance.toJSON());
+model.evaluate(instance, "play");
+model.evaluate(instance, "pause");
+model.evaluate(instance, "stop");
+// console.log(instance.toJSON());